Modify signup/singin/signout into one service..Test cases passed for UserContoller
