--- quora-service/src/main/java/com/upgrad/quora/service/business/SigninAuthenticationService.java
@@ -1,58 +1,0 @@
-package com.upgrad.quora.service.business;
-
-import com.upgrad.quora.service.dao.UserDao;
-import com.upgrad.quora.service.entity.UserAuthTokenEntity;
-import com.upgrad.quora.service.entity.UserEntity;
-import com.upgrad.quora.service.exception.AuthenticationFailedException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.ZonedDateTime;
-import java.util.UUID;
-
-@Service
-public class SigninAuthenticationService {
-
-    @Autowired
-    private UserDao userDao;
-
-    @Autowired
-    private PasswordCryptographyProvider cryptographyProvider;
-
-    @Transactional(propagation = Propagation.REQUIRED)
-    public UserAuthTokenEntity authenticate(final String username,final String password) throws AuthenticationFailedException {
-
-        UserEntity user = userDao.userByUsername(username);
-
-        if(user == null)
-            throw new AuthenticationFailedException("ATH-001","This username does not exist");
-
-        final String encryptedPassword = PasswordCryptographyProvider.encrypt(password, user.getSalt());
-
-        if(encryptedPassword.equals(user.getPassword())){
-
-            JwtTokenProvider jwtTokenProvider = new JwtTokenProvider(encryptedPassword);
-            UserAuthTokenEntity userAuthTokenEntity = new UserAuthTokenEntity();
-            userAuthTokenEntity.setUuid(UUID.randomUUID().toString());
-            userAuthTokenEntity.setUser(user);
-
-            final ZonedDateTime now = ZonedDateTime.now();
-            final ZonedDateTime expiresAt = now.plusHours(8);
-
-            userAuthTokenEntity.setAccessToken(jwtTokenProvider.generateToken(user.getUuid(),now,expiresAt));
-            userAuthTokenEntity.setLoginAt(now);
-            userAuthTokenEntity.setExpiresAt(expiresAt);
-
-            userDao.createAuthToken(userAuthTokenEntity);
-            userDao.updateUser(user);
-
-            return userAuthTokenEntity;
-
-        }else {
-            throw new AuthenticationFailedException("ATH-002","Password failed");
-        }
-
-    }
-}
