--- quora-service/src/main/java/com/upgrad/quora/service/business/SignupBusinessService.java
@@ -1,48 +1,0 @@
-package com.upgrad.quora.service.business;
-
-import com.upgrad.quora.service.dao.UserDao;
-import com.upgrad.quora.service.entity.UserEntity;
-import com.upgrad.quora.service.exception.SignUpRestrictedException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class SignupBusinessService {
-
-    @Autowired
-    private UserDao userDao;
-
-    @Autowired
-    private PasswordCryptographyProvider cryptographyProvider;
-
-    @Transactional(propagation = Propagation.REQUIRED)
-    public UserEntity signUp(final UserEntity userEntity) throws SignUpRestrictedException {
-
-        if(checkExistingUsername(userEntity.getUsername())) {
-            throw new SignUpRestrictedException("SGR-001", "Try any other Username, this Username has already been taken");
-        }
-
-        if(checkExistingEmail(userEntity.getEmail())) {
-            throw new SignUpRestrictedException("SGR-002", "This user has already been registered, try with any other emailId");
-        }
-
-        final String password = userEntity.getPassword();
-
-        final String[] encryptedText = cryptographyProvider.encrypt(password);
-
-        userEntity.setSalt(encryptedText[0]);
-        userEntity.setPassword(encryptedText[1]);
-
-        return userDao.signUp(userEntity);
-    }
-
-    public boolean checkExistingUsername(final String username){
-        return userDao.userByUsername(username) != null;
-    }
-
-    public boolean checkExistingEmail(final String email){
-        return userDao.userByEmail(email) != null;
-    }
-}
