+++ quora-service/src/main/java/com/upgrad/quora/service/dao/UserDao.java
@@ -64,4 +64,5 @@ public class UserDao {
     public void updateUserAuth(UserAuthTokenEntity userAuthTokenEntity){
         entityManager.merge(userAuthTokenEntity);
     }
+
 }
